+++ internal/tool/builtins.go
@@ -1,0 +1,133 @@
+package tool
+
+import (
+	"context"
+	"errors"
+	"fmt"
+	"net"
+	"time"
+
+	"github.com/marcodenic/agentry/internal/patch"
+	"github.com/marcodenic/agentry/internal/team"
+)
+
+// builtinSpec defines builtin schema and execution.
+type builtinSpec struct {
+	Desc   string
+	Schema map[string]any
+	Exec   ExecFn
+}
+
+// builtinMap holds safe builtin tools keyed by name.
+var builtinMap = map[string]builtinSpec{
+	"echo": {
+		Desc: "Repeat a string",
+		Schema: map[string]any{
+			"type":       "object",
+			"properties": map[string]any{"text": map[string]any{"type": "string"}},
+			"required":   []string{"text"},
+		},
+		Exec: func(ctx context.Context, args map[string]any) (string, error) {
+			txt, _ := args["text"].(string)
+			return txt, nil
+		},
+	},
+	"ping": {
+		Desc: "Ping a host",
+		Schema: map[string]any{
+			"type":       "object",
+			"properties": map[string]any{"host": map[string]any{"type": "string"}},
+			"required":   []string{"host"},
+			"example":    map[string]any{"host": "example.com"},
+		},
+		Exec: func(ctx context.Context, args map[string]any) (string, error) {
+			host, _ := args["host"].(string)
+			if host == "" {
+				return "", errors.New("missing host")
+			}
+			d := net.Dialer{Timeout: 3 * time.Second}
+			start := time.Now()
+			conn, err := d.DialContext(ctx, "tcp", host+":80")
+			if err != nil {
+				return "", err
+			}
+			_ = conn.Close()
+			return fmt.Sprintf("pong in %v", time.Since(start)), nil
+		},
+	},
+	"mcp": {
+		Desc: "Execute an MCP command",
+		Schema: map[string]any{
+			"type": "object",
+			"properties": map[string]any{
+				"host":    map[string]any{"type": "string"},
+				"port":    map[string]any{"type": "integer"},
+				"command": map[string]any{"type": "string"},
+			},
+			"required": []string{"host", "port", "command"},
+			"example":  map[string]any{"host": "localhost", "port": 1234, "command": "hello"},
+		},
+		Exec: func(ctx context.Context, args map[string]any) (string, error) {
+			host, _ := args["host"].(string)
+			port, _ := args["port"].(float64)
+			cmd, _ := args["command"].(string)
+			addr := fmt.Sprintf("%s:%d", host, int(port))
+			conn, err := net.DialTimeout("tcp", addr, 3*time.Second)
+			if err != nil {
+				return "", err
+			}
+			defer conn.Close()
+			_, _ = conn.Write([]byte(cmd + "\n"))
+			buf := make([]byte, 1024)
+			n, _ := conn.Read(buf)
+			return string(buf[:n]), nil
+		},
+	},
+	"agent-call": {
+		Desc: "Delegate to another agent",
+		Schema: map[string]any{
+			"type": "object",
+			"properties": map[string]any{
+				"agent": map[string]any{"type": "string"},
+				"input": map[string]any{"type": "string"},
+			},
+			"required": []string{"agent", "input"},
+			"example": map[string]any{
+				"agent": "Agent1",
+				"input": "Hello, how are you?",
+			},
+		},
+		Exec: func(ctx context.Context, args map[string]any) (string, error) {
+			name, _ := args["agent"].(string)
+			input, _ := args["input"].(string)
+			t, ok := team.FromContext(ctx)
+			if !ok {
+				return "", errors.New("team not found in context")
+			}
+			return t.Call(ctx, name, input)
+		},
+	},
+}
+
+func init() {
+	builtinMap["patch"] = builtinSpec{
+		Desc: "Apply a unified diff patch",
+		Schema: map[string]any{
+			"type":       "object",
+			"properties": map[string]any{"patch": map[string]any{"type": "string"}},
+			"required":   []string{"patch"},
+			"example":    map[string]any{"patch": ""},
+		},
+		Exec: func(ctx context.Context, args map[string]any) (string, error) {
+			patchStr, _ := args["patch"].(string)
+			if patchStr == "" {
+				return "", errors.New("missing patch")
+			}
+			res, err := patch.Apply(patchStr)
+			if err != nil {
+				return "", err
+			}
+			return patch.MarshalResult(res)
+		},
+	}
+}
