+++ internal/tool/manifest.go
@@ -8,9 +8,7 @@ import (
 	"errors"
 	"fmt"
 	"io"
-	"net"
 	"net/http"
-	"net/url"
 	"os"
 	"os/exec"
 	"path/filepath"
@@ -20,8 +18,6 @@ import (
 	"time"
 
 	"github.com/marcodenic/agentry/internal/config"
-	"github.com/marcodenic/agentry/internal/patch"
-	"github.com/marcodenic/agentry/internal/team"
 	"github.com/marcodenic/agentry/pkg/sbox"
 )
 
@@ -123,534 +119,6 @@ func checkForOverwrite(path string) error {
 		return fmt.Errorf("file %s changed since viewed", path)
 	}
 	return nil
-}
-
-type Tool interface {
-	Name() string
-	Description() string
-	JSONSchema() map[string]any
-	Execute(ctx context.Context, args map[string]any) (string, error)
-}
-
-type simpleTool struct {
-	name    string
-	desc    string
-	schema  map[string]any
-	fn      func(context.Context, map[string]any) (string, error)
-	allowed bool
-}
-
-func New(name, desc string, fn func(context.Context, map[string]any) (string, error)) Tool {
-	return &simpleTool{name: name, desc: desc, fn: fn, schema: map[string]any{"type": "object"}, allowed: true}
-}
-
-func NewWithSchema(name, desc string, schema map[string]any, fn func(context.Context, map[string]any) (string, error)) Tool {
-	return &simpleTool{name: name, desc: desc, fn: fn, schema: schema, allowed: true}
-}
-
-func (t *simpleTool) Name() string               { return t.name }
-func (t *simpleTool) Description() string        { return t.desc }
-func (t *simpleTool) JSONSchema() map[string]any { return t.schema }
-func (t *simpleTool) Execute(ctx context.Context, args map[string]any) (string, error) {
-	if !t.allowed {
-		return "", fmt.Errorf("%w: %s", ErrToolDenied, t.name)
-	}
-	if !permitted(t.name) {
-		return "", fmt.Errorf("%w: %s", ErrToolDenied, t.name)
-	}
-	return t.fn(ctx, args)
-}
-
-type Registry map[string]Tool
-
-func (r Registry) Use(name string) (Tool, bool) {
-	t, ok := r[name]
-	return t, ok
-}
-
-// ExecFn defines the signature for tool execution functions.
-type ExecFn func(context.Context, map[string]any) (string, error)
-
-// builtinSpec defines builtin schema and execution.
-type builtinSpec struct {
-	Desc   string
-	Schema map[string]any
-	Exec   ExecFn
-}
-
-// builtinMap holds safe builtin tools keyed by name.
-var builtinMap = map[string]builtinSpec{
-	"echo": {
-		Desc: "Repeat a string",
-		Schema: map[string]any{
-			"type":       "object",
-			"properties": map[string]any{"text": map[string]any{"type": "string"}},
-			"required":   []string{"text"},
-		},
-		Exec: func(ctx context.Context, args map[string]any) (string, error) {
-			txt, _ := args["text"].(string)
-			return txt, nil
-		},
-	},
-	"ping": {
-		Desc: "Ping a host",
-		Schema: map[string]any{
-			"type":       "object",
-			"properties": map[string]any{"host": map[string]any{"type": "string"}},
-			"required":   []string{"host"},
-			"example":    map[string]any{"host": "example.com"},
-		},
-		Exec: func(ctx context.Context, args map[string]any) (string, error) {
-			host, _ := args["host"].(string)
-			if host == "" {
-				return "", errors.New("missing host")
-			}
-			d := net.Dialer{Timeout: 3 * time.Second}
-			start := time.Now()
-			conn, err := d.DialContext(ctx, "tcp", host+":80")
-			if err != nil {
-				return "", err
-			}
-			_ = conn.Close()
-			return fmt.Sprintf("pong in %v", time.Since(start)), nil
-		},
-	},
-	"mcp": {
-		Desc: "Execute an MCP command",
-		Schema: map[string]any{
-			"type": "object",
-			"properties": map[string]any{
-				"host":    map[string]any{"type": "string"},
-				"port":    map[string]any{"type": "integer"},
-				"command": map[string]any{"type": "string"},
-			},
-			"required": []string{"host", "port", "command"},
-			"example":  map[string]any{"host": "localhost", "port": 1234, "command": "hello"},
-		},
-		Exec: func(ctx context.Context, args map[string]any) (string, error) {
-			host, _ := args["host"].(string)
-			var portStr string
-			switch p := args["port"].(type) {
-			case float64:
-				portStr = fmt.Sprintf("%d", int(p))
-			case int:
-				portStr = fmt.Sprintf("%d", p)
-			case string:
-				portStr = p
-			}
-			cmdStr, _ := args["command"].(string)
-			if host == "" || portStr == "" || cmdStr == "" {
-				return "", errors.New("missing args")
-			}
-			d := net.Dialer{Timeout: 3 * time.Second}
-			conn, err := d.DialContext(ctx, "tcp", net.JoinHostPort(host, portStr))
-			if err != nil {
-				return "", err
-			}
-			defer conn.Close()
-			_ = conn.SetDeadline(time.Now().Add(3 * time.Second))
-			if _, err := conn.Write([]byte(cmdStr)); err != nil {
-				return "", err
-			}
-			resp, err := io.ReadAll(conn)
-			if err != nil {
-				return "", err
-			}
-			return string(resp), nil
-		},
-	},
-	"bash": {
-		Desc: "Execute a bash command",
-		Schema: map[string]any{
-			"type":       "object",
-			"properties": map[string]any{"cmd": map[string]any{"type": "string"}},
-			"required":   []string{"cmd"},
-			"example":    map[string]any{"cmd": "echo hi"},
-		}, Exec: func(ctx context.Context, args map[string]any) (string, error) {
-			cmdStr, _ := args["cmd"].(string)
-			if cmdStr == "" {
-				return "", errors.New("missing cmd")
-			}
-			if _, err := exec.LookPath(shellCmd); err != nil {
-				return "", fmt.Errorf("%s not found – install or use another tool", shellCmd)
-			}
-			cmd := exec.CommandContext(ctx, shellCmd, shellFlag, cmdStr)
-			out, err := cmd.CombinedOutput()
-			return string(out), err
-		},
-	},
-	"branch-tidy": {
-		Desc: "Delete all local Git branches except the current one",
-		Schema: map[string]any{
-			"type": "object",
-			"properties": map[string]any{
-				"force": map[string]any{
-					"type":        "boolean",
-					"description": "Use -D flag to force delete branches (default: false uses -d)",
-					"default":     false,
-				},
-				"dry-run": map[string]any{
-					"type":        "boolean",
-					"description": "Show which branches would be deleted without actually deleting them",
-					"default":     false,
-				},
-			},
-			"example": map[string]any{"force": false, "dry-run": true},
-		},
-		Exec: func(ctx context.Context, args map[string]any) (string, error) {
-			force, _ := args["force"].(bool)
-			dryRun, _ := args["dry-run"].(bool)
-
-			// Get current branch
-			currentCmd := exec.CommandContext(ctx, "git", "rev-parse", "--abbrev-ref", "HEAD")
-			currentOut, err := currentCmd.Output()
-			if err != nil {
-				return "", fmt.Errorf("failed to get current branch: %v", err)
-			}
-			current := strings.TrimSpace(string(currentOut))
-			if current == "" {
-				return "", errors.New("could not determine current branch")
-			}
-
-			// Get all local branches
-			branchCmd := exec.CommandContext(ctx, "git", "branch")
-			branchOut, err := branchCmd.Output()
-			if err != nil {
-				return "", fmt.Errorf("failed to list branches: %v", err)
-			}
-
-			var result strings.Builder
-			var deleted []string
-			var skipped []string
-
-			lines := strings.Split(string(branchOut), "\n")
-			for _, line := range lines {
-				line = strings.TrimSpace(line)
-				if line == "" {
-					continue
-				}
-
-				// Skip current branch (marked with * or just the name)
-				branch := strings.TrimPrefix(line, "* ")
-				if branch == current || strings.HasPrefix(line, "* ") {
-					skipped = append(skipped, branch)
-					continue
-				}
-
-				// Skip branches that match common protected patterns
-				if branch == "main" || branch == "master" || branch == "develop" || branch == "development" {
-					skipped = append(skipped, branch)
-					continue
-				}
-
-				if dryRun {
-					deleted = append(deleted, branch)
-				} else {
-					// Delete the branch
-					deleteFlag := "-d"
-					if force {
-						deleteFlag = "-D"
-					}
-
-					deleteCmd := exec.CommandContext(ctx, "git", "branch", deleteFlag, branch)
-					if err := deleteCmd.Run(); err != nil {
-						result.WriteString(fmt.Sprintf("Failed to delete branch '%s': %v\n", branch, err))
-					} else {
-						deleted = append(deleted, branch)
-					}
-				}
-			}
-
-			// Build result message
-			if dryRun {
-				result.WriteString("DRY RUN - branches that would be deleted:\n")
-				for _, branch := range deleted {
-					result.WriteString(fmt.Sprintf("  - %s\n", branch))
-				}
-			} else {
-				result.WriteString(fmt.Sprintf("Successfully deleted %d branches:\n", len(deleted)))
-				for _, branch := range deleted {
-					result.WriteString(fmt.Sprintf("  - %s\n", branch))
-				}
-			}
-
-			if len(skipped) > 0 {
-				result.WriteString("\nSkipped branches:\n")
-				for _, branch := range skipped {
-					result.WriteString(fmt.Sprintf("  - %s (protected or current)\n", branch))
-				}
-			}
-
-			if len(deleted) == 0 && !dryRun {
-				result.WriteString("No branches to delete.")
-			}
-
-			return result.String(), nil
-		},
-	},
-	"fetch": {
-		Desc: "Download content from a URL",
-		Schema: map[string]any{
-			"type":       "object",
-			"properties": map[string]any{"url": map[string]any{"type": "string"}},
-			"required":   []string{"url"},
-			"example":    map[string]any{"url": "https://example.com"},
-		},
-		Exec: func(ctx context.Context, args map[string]any) (string, error) {
-			u, _ := args["url"].(string)
-			if u == "" {
-				return "", errors.New("missing url")
-			}
-			req, err := http.NewRequestWithContext(ctx, http.MethodGet, u, nil)
-			if err != nil {
-				return "", err
-			}
-			resp, err := http.DefaultClient.Do(req)
-			if err != nil {
-				return "", err
-			}
-			defer resp.Body.Close()
-			b, err := io.ReadAll(resp.Body)
-			if err != nil {
-				return "", err
-			}
-			return string(b), nil
-		},
-	},
-	"ls": {
-		Desc: "List directory contents",
-		Schema: map[string]any{
-			"type":       "object",
-			"properties": map[string]any{"path": map[string]any{"type": "string"}},
-			"example":    map[string]any{"path": "."},
-		},
-		Exec: func(ctx context.Context, args map[string]any) (string, error) {
-			path, _ := args["path"].(string)
-			if path == "" {
-				path = "."
-			}
-			path = absPath(path)
-			entries, err := os.ReadDir(path)
-			if err != nil {
-				return "", err
-			}
-			names := make([]string, len(entries))
-			for i, e := range entries {
-				names[i] = e.Name()
-			}
-			return strings.Join(names, "\n"), nil
-		},
-	},
-	"view": {
-		Desc: "Read a file",
-		Schema: map[string]any{
-			"type":       "object",
-			"properties": map[string]any{"path": map[string]any{"type": "string"}},
-			"required":   []string{"path"},
-			"example":    map[string]any{"path": "go.mod"},
-		},
-		Exec: func(ctx context.Context, args map[string]any) (string, error) {
-			path, _ := args["path"].(string)
-			if path == "" {
-				return "", errors.New("missing path")
-			}
-			path = absPath(path)
-			if _, err := os.Stat(path); errors.Is(err, os.ErrNotExist) {
-				return "", fmt.Errorf("file %s does not exist – create it first", path)
-			}
-			b, err := os.ReadFile(path)
-			if err != nil {
-				return "", err
-			}
-			if err := recordView(path); err != nil {
-				return "", err
-			}
-			return string(b), nil
-		},
-	},
-	"write": {
-		Desc: "Create or overwrite a file",
-		Schema: map[string]any{
-			"type": "object",
-			"properties": map[string]any{
-				"path": map[string]any{"type": "string"},
-				"text": map[string]any{"type": "string"},
-			},
-			"required": []string{"path", "text"},
-			"example":  map[string]any{"path": "tmp.txt", "text": "hi"},
-		},
-		Exec: func(ctx context.Context, args map[string]any) (string, error) {
-			path, _ := args["path"].(string)
-			text, _ := args["text"].(string)
-			if path == "" {
-				return "", errors.New("missing path")
-			}
-			path = absPath(path)
-			if err := checkForOverwrite(path); err != nil {
-				return "", err
-			}
-			if err := os.WriteFile(path, []byte(text), 0644); err != nil {
-				return "", err
-			}
-			_ = recordView(path)
-			return "written", nil
-		},
-	},
-	"glob": {
-		Desc: "Find files by pattern",
-		Schema: map[string]any{
-			"type":       "object",
-			"properties": map[string]any{"pattern": map[string]any{"type": "string"}},
-			"required":   []string{"pattern"},
-			"example":    map[string]any{"pattern": "*.go"},
-		},
-		Exec: func(ctx context.Context, args map[string]any) (string, error) {
-			pattern, _ := args["pattern"].(string)
-			if pattern == "" {
-				return "", errors.New("missing pattern")
-			}
-			files, err := filepath.Glob(pattern)
-			if err != nil {
-				return "", err
-			}
-			return strings.Join(files, "\n"), nil
-		},
-	},
-	"grep": {
-		Desc: "Search file contents",
-		Schema: map[string]any{
-			"type": "object",
-			"properties": map[string]any{
-				"pattern": map[string]any{"type": "string"},
-				"path":    map[string]any{"type": "string"},
-			},
-			"required": []string{"pattern", "path"},
-			"example":  map[string]any{"pattern": "hello", "path": "go.mod"},
-		},
-		Exec: func(ctx context.Context, args map[string]any) (string, error) {
-			pattern, _ := args["pattern"].(string)
-			path, _ := args["path"].(string)
-			if path == "" || pattern == "" {
-				return "", errors.New("missing args")
-			}
-			path = absPath(path)
-			if _, err := os.Stat(path); errors.Is(err, os.ErrNotExist) {
-				return "", fmt.Errorf("file %s does not exist – create it first", path)
-			}
-			b, err := os.ReadFile(path)
-			if err != nil {
-				return "", err
-			}
-			lines := strings.Split(string(b), "\n")
-			var out []string
-			for _, l := range lines {
-				if strings.Contains(l, pattern) {
-					out = append(out, l)
-				}
-			}
-			return strings.Join(out, "\n"), nil
-		},
-	},
-	"edit": {
-		Desc: "Update an existing file",
-		Schema: map[string]any{
-			"type": "object",
-			"properties": map[string]any{
-				"path": map[string]any{"type": "string"},
-				"text": map[string]any{"type": "string"},
-			},
-			"required": []string{"path", "text"},
-			"example":  map[string]any{"path": "tmp.txt", "text": "bye"},
-		},
-		Exec: func(ctx context.Context, args map[string]any) (string, error) {
-			path, _ := args["path"].(string)
-			text, _ := args["text"].(string)
-			if path == "" {
-				return "", errors.New("missing path")
-			}
-			path = absPath(path)
-			if err := checkForOverwrite(path); err != nil {
-				return "", err
-			}
-			if err := os.WriteFile(path, []byte(text), 0644); err != nil {
-				return "", err
-			}
-			_ = recordView(path)
-			return "edited", nil
-		},
-	},
-	"sourcegraph": {
-		Desc: "Search public repositories",
-		Schema: map[string]any{
-			"type":       "object",
-			"properties": map[string]any{"query": map[string]any{"type": "string"}},
-			"required":   []string{"query"},
-			"example":    map[string]any{"query": "grpc"},
-		},
-		Exec: func(ctx context.Context, args map[string]any) (string, error) {
-			q, _ := args["query"].(string)
-			if q == "" {
-				return "", errors.New("missing query")
-			}
-			url := "https://sourcegraph.com/search?q=" + url.QueryEscape(q) + "&format=json"
-			resp, err := http.Get(url)
-			if err != nil {
-				return "", err
-			}
-			defer resp.Body.Close()
-			b, err := io.ReadAll(resp.Body)
-			if err != nil {
-				return "", err
-			}
-			return string(b), nil
-		},
-	}, "agent": {
-		Desc: "Send a message to another agent",
-		Schema: map[string]any{
-			"type": "object",
-			"properties": map[string]any{
-				"agent": map[string]any{"type": "string"},
-				"input": map[string]any{"type": "string"},
-			},
-			"required": []string{"agent", "input"},
-			"example": map[string]any{
-				"agent": "Agent1",
-				"input": "Hello, how are you?",
-			},
-		}, Exec: func(ctx context.Context, args map[string]any) (string, error) {
-			name, _ := args["agent"].(string)
-			input, _ := args["input"].(string)
-
-			// Look for team interface in context
-			t, ok := team.FromContext(ctx)
-			if !ok {				return "", errors.New("team not found in context")
-			}
-			return t.Call(ctx, name, input)
-		},
-	},
-}
-
-func init() {
-	builtinMap["patch"] = builtinSpec{
-		Desc: "Apply a unified diff patch",
-		Schema: map[string]any{
-			"type":       "object",
-			"properties": map[string]any{"patch": map[string]any{"type": "string"}},
-			"required":   []string{"patch"},
-			"example":    map[string]any{"patch": ""},
-		},
-		Exec: func(ctx context.Context, args map[string]any) (string, error) {
-			patchStr, _ := args["patch"].(string)
-			if patchStr == "" {
-				return "", errors.New("missing patch")
-			}
-			res, err := patch.Apply(patchStr)
-			if err != nil {
-				return "", err
-			}
-			return patch.MarshalResult(res)
-		},
-	}
 }
 
 // DefaultRegistry returns all builtin tools.
