+++ internal/tui/commands.go
@@ -1,0 +1,211 @@
+package tui
+
+import (
+	"bufio"
+	"context"
+	"io"
+	"strconv"
+	"strings"
+	"time"
+
+	"github.com/charmbracelet/bubbles/spinner"
+	tea "github.com/charmbracelet/bubbletea"
+	"github.com/charmbracelet/lipgloss"
+	"github.com/google/uuid"
+	"github.com/marcodenic/agentry/internal/teamctx"
+	"github.com/marcodenic/agentry/internal/trace"
+)
+
+// startAgent runs an agent with the given input and streams its output.
+func (m Model) startAgent(id uuid.UUID, input string) (Model, tea.Cmd) {
+	info := m.infos[id]
+	info.Status = StatusRunning
+	info.Spinner = spinner.New()
+	info.Spinner.Spinner = spinner.Line
+	info.Spinner.Style = lipgloss.NewStyle().Foreground(lipgloss.Color(m.theme.AIBarColor))
+	info.History += m.userBar() + " " + input + "\n"
+	base := lipgloss.NewStyle().Foreground(lipgloss.Color(m.theme.Palette.Foreground)).Background(lipgloss.Color(m.theme.Palette.Background))
+	m.vp.SetContent(base.Copy().Width(m.vp.Width).Render(info.History))
+
+	pr, pw := io.Pipe()
+	errCh := make(chan error, 1)
+	tracer := trace.NewJSONL(pw)
+	if info.Agent.Tracer != nil {
+		info.Agent.Tracer = trace.NewMulti(info.Agent.Tracer, tracer)
+	} else {
+		info.Agent.Tracer = tracer
+	}
+	info.Scanner = bufio.NewScanner(pr)
+	ctx := context.WithValue(context.Background(), teamctx.Key{}, m.team)
+	ctx, cancel := context.WithCancel(ctx)
+	info.Cancel = cancel
+	m.infos[id] = info
+	go func() {
+		_, err := info.Agent.Run(ctx, input)
+		pw.Close()
+		errCh <- err
+	}()
+	m.infos[id] = info
+	return m, tea.Batch(m.readCmd(id), waitErr(errCh), info.Spinner.Tick)
+}
+
+// handleCommand parses a slash command and dispatches to the appropriate handler.
+func (m Model) handleCommand(cmd string) (Model, tea.Cmd) {
+	fields := strings.Fields(cmd)
+	if len(fields) == 0 {
+		return m, nil
+	}
+	switch fields[0] {
+	case "/spawn":
+		return m.handleSpawn(fields[1:])
+	case "/switch":
+		return m.handleSwitch(fields[1:])
+	case "/stop":
+		return m.handleStop(fields[1:])
+		return m.handleConverse(fields[1:])
+	default:
+		return m, nil
+	}
+}
+
+// handleSpawn creates a new agent and adds it to the panel.
+func (m Model) handleSpawn(args []string) (Model, tea.Cmd) {
+	name := "agent"
+	role := ""
+	if len(args) > 0 {
+		name = args[0]
+	}
+	if len(args) > 1 {
+		role = args[1]
+	}
+	if len(m.agents) == 0 {
+		return m, nil
+	}
+	ag := m.agents[0].Spawn()
+	sp := spinner.New()
+	sp.Spinner = spinner.Line
+	sp.Style = lipgloss.NewStyle().Foreground(lipgloss.Color(m.theme.AIBarColor))
+	info := &AgentInfo{
+		Agent:           ag,
+		Status:          StatusIdle,
+		Spinner:         sp,
+		Name:            name,
+		Role:            role,
+		ActivityData:    make([]float64, 0),
+		ActivityTimes:   make([]time.Time, 0),
+		CurrentActivity: 0,
+		LastActivity:    time.Time{},
+		TokenHistory:    []int{},
+	}
+	m.infos[ag.ID] = info
+	m.order = append(m.order, ag.ID)
+	if m.team != nil {
+		m.team.Add(name, ag)
+	}
+	m.active = ag.ID
+	m.vp.SetContent("")
+	return m, nil
+}
+
+// handleSwitch focuses the agent whose ID prefix matches the argument.
+func (m Model) handleSwitch(args []string) (Model, tea.Cmd) {
+	if len(args) == 0 {
+		return m, nil
+	}
+	prefix := args[0]
+	for _, id := range m.order {
+		if strings.HasPrefix(id.String(), prefix) {
+			m.active = id
+			if info, ok := m.infos[id]; ok {
+				base := lipgloss.NewStyle().Foreground(lipgloss.Color(m.theme.Palette.Foreground)).Background(lipgloss.Color(m.theme.Palette.Background))
+				m.vp.SetContent(base.Copy().Width(m.vp.Width).Render(info.History))
+			}
+			break
+		}
+	}
+	return m, nil
+}
+
+// handleStop cancels a running agent.
+func (m Model) handleStop(args []string) (Model, tea.Cmd) {
+	id := m.active
+	if len(args) > 0 {
+		pref := args[0]
+		for _, aid := range m.order {
+			if strings.HasPrefix(aid.String(), pref) {
+				id = aid
+				break
+			}
+		}
+	}
+	if info, ok := m.infos[id]; ok {
+		if info.Cancel != nil {
+			info.Cancel()
+		}
+		info.Status = StatusStopped
+		m.infos[id] = info
+	}
+	return m, nil
+}
+
+// handleConverse launches a temporary team conversation.
+func (m Model) handleConverse(args []string) (Model, tea.Cmd) {
+	if len(args) < 2 {
+		return m, nil
+	}
+	n, err := strconv.Atoi(args[0])
+	if err != nil {
+		return m, nil
+	}
+	topic := strings.Join(args[1:], " ")
+	if len(m.agents) == 0 {
+		return m, nil
+	}
+	tm, err := NewTeam(m.agents[0], n, topic)
+	if err != nil {
+		m.err = err
+		return m, nil
+	}
+	go func() { _ = tea.NewProgram(tm, tea.WithAltScreen(), tea.WithMouseCellMotion()).Start() }()
+	return m, nil
+}
+
+// cycleActive moves the focus to the next or previous agent.
+func (m Model) cycleActive(delta int) Model {
+	if len(m.order) == 0 {
+		return m
+	}
+	idx := 0
+	for i, id := range m.order {
+		if id == m.active {
+			idx = i
+			break
+		}
+	}
+	idx = (idx + delta + len(m.order)) % len(m.order)
+	m.active = m.order[idx]
+	if info, ok := m.infos[m.active]; ok {
+		base := lipgloss.NewStyle().Foreground(lipgloss.Color(m.theme.Palette.Foreground)).Background(lipgloss.Color(m.theme.Palette.Background))
+		m.vp.SetContent(base.Copy().Width(m.vp.Width).Render(info.History))
+	}
+	return m
+}
+
+// jumpToAgent sets the active agent by index.
+func (m Model) jumpToAgent(index int) Model {
+	if len(m.order) == 0 {
+		return m
+	}
+	if index < 0 {
+		index = 0
+	}
+	if index >= len(m.order) {
+		index = len(m.order) - 1
+	}
+	m.active = m.order[index]
+	if info, ok := m.infos[m.active]; ok {
+		base := lipgloss.NewStyle().Foreground(lipgloss.Color(m.theme.Palette.Foreground)).Background(lipgloss.Color(m.theme.Palette.Background))
+		m.vp.SetContent(base.Copy().Width(m.vp.Width).Render(info.History))
+	}
+	return m
+}
